requirements.txt
streamlit
speechrecognition
gtts
pygame
PyMuPDF
google-generativeai
fitz
tempfile
io
